requirements.txt
anthropic>=0.95.0
fastapi>=0.124.0
uvicorn[standard]>=0.38.0
chromadb>=1.5.7
sentence-transformers>=5.4.1
python-dotenv>=1.2.2
pydantic>=2.0.0
